save progress refactor
