+++ ssr/src/page/src/upload/ai/components/mod.rs
@@ -1,0 +1,11 @@
+pub mod model_dropdown;
+pub mod loading_screen;
+pub mod result_screen;
+pub mod pre_upload;
+pub mod page;
+
+pub use model_dropdown::ModelDropdown;
+pub use loading_screen::VideoGenerationLoadingScreen;
+pub use result_screen::VideoResultScreen;
+pub use pre_upload::PreUploadAiView;
+pub use page::UploadAiPostPage;
