+++ ssr/src/page/src/upload/ai/helpers.rs
@@ -1,0 +1,26 @@
+use candid::Principal;
+use super::models::VideoModel;
+
+// Helper function to generate video using videogen.rs
+pub async fn generate_video(
+    _user_principal: Principal,
+    prompt: String,
+    _model: VideoModel,
+    _image_data: Option<String>,
+) -> Result<String, Box<dyn std::error::Error>> {
+    // For now, simulate video generation with a delay
+    // TODO: Implement actual videogen.rs integration
+    leptos::logging::log!("Starting video generation with prompt: {}", prompt);
+
+    // Simulate some processing time
+    #[cfg(feature = "hydrate")]
+    {
+        gloo::timers::future::TimeoutFuture::new(2000).await;
+    }
+
+    // Mock successful response
+    let mock_video_url = "https://storage.googleapis.com/yral_ai_generated_videos/veo-output/5790230970440583959/sample_0.mp4";
+
+    leptos::logging::log!("Video generation completed with URL: {}", mock_video_url);
+    Ok(mock_video_url.to_string())
+}
