+++ ssr/src/page/src/upload/ai/mod.rs
@@ -1,0 +1,8 @@
+pub mod types;
+pub mod models;
+pub mod helpers;
+pub mod server;
+pub mod components;
+
+// Re-export the main page component for ease of use
+pub use components::UploadAiPostPage;
